Data for Example 10.2 from Generalized Linear Mixed Models: Modern Concepts, Methods and Applications by Stroup, Ptukhina, and Garai (2024, 2nd ed.)

DataSet10.2 is used for Two-way nested random effects model and BLUP estimation

Usage

data(DataSet10.2)

Format

A data.frame with 28 rows and 3 variables with levels of b nested within levels of a.

Details

a is a factor with 7 levels

b is a factor with 2 levels

y is a Gaussian response variable

References

Stroup, W. W., Ptukhina, M., and Garai, S. (2024). Generalized Linear Mixed Models: Modern Concepts, Methods and Applications (2nd ed.). CRC Press.

See also

Author

Muhammad Yaseen (myaseen208@gmail.com) Adeela Munawar (adeela.uaf@gmail.com)

Examples

data(DataSet10.2)